requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.41.1
requests==2.32.3
plotly==5.24.1
